+++ chaosmagpy/__init__.py
@@ -10,4 +10,4 @@ __all__ = [
     "basicConfig", "synth_values"
 ]
 
-__version__ = "0.8"
+__version__ = "0.9-dev"